misc fixes with checking syscalls returns and sign-mismatches (USC)   don't corrupt hostnames when checking torque.cfg SERVERHOST names (USC)   don't corrupt memory when parsing env vars without an '=' (USC)

+++ src/lib/Libdis/disrcs.c
@@ -131,7 +131,7 @@ char *disrcs(stream, nchars, retval)
 		        	locret = DIS_NOMALLOC;
 			else {
 				if ((*dis_gets)(stream, value,
-						(size_t)count) != (size_t)count)
+						(size_t)count) != (int)count)
 			        	locret = DIS_PROTO;
 				else
 					value[count] = '\0';
