+++ src/lib/Libdis/disrf.c
@@ -234,7 +234,7 @@ static int disrd_(
 
       if (count > 1) 
         {
-        if ((*dis_gets)(stream,dis_buffer + 1,count - 1) != count - 1)
+        if ((*dis_gets)(stream,dis_buffer + 1,count - 1) != (int)count - 1)
           {
           return(DIS_EOD);
           }
